requirements.txt
pytest==7.4.1
allure-pytest==2.13.2
PyHamcrest==2.0.4
requests==2.31.0
